+++ internal/data/mysql/store_mysql_integration_test.go
@@ -11,8 +11,8 @@ import (
 
 	"github.com/JekYUlll/Dipole/db/migrations"
 	"github.com/JekYUlll/Dipole/internal/data/migration"
-	"github.com/JekYUlll/Dipole/internal/data/mysql/generated"
-	"github.com/JekYUlll/Dipole/internal/data/mysql/mapper"
+	"github.com/JekYUlll/Dipole/internal/platform/mysql/generated"
+	"github.com/JekYUlll/Dipole/internal/platform/mysql/mapper"
 	"github.com/JekYUlll/Dipole/internal/model"
 	mysqlStore "github.com/JekYUlll/Dipole/internal/platform/mysql"
 	mysqlDriver "github.com/go-sql-driver/mysql"